+++ common/src/main/java/com/fej1fun/potentials/providers/FluidProvider.java
@@ -12,6 +12,9 @@ public class FluidProvider {
     public interface BLOCK {
         @Nullable UniversalFluidStorage getFluidTank(@Nullable Direction direction);
     }
+    public interface ENTITY {
+        @Nullable UniversalFluidStorage getFluidTank(@Nullable Direction direction);
+    }
     public interface ITEM {
         @Nullable UniversalFluidItemStorage getFluidTank(@NotNull ItemStack stack);
     }